requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-extensions==3.2.3
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==22.0.0
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.11.0
whitenoise==6.6.0
